added build step to docker-compose and renamed pp-pairing
