+++ backend/services/pairing/pairing-back/src/models/linked-list.ts
