+++ backend/services/pairing/pairing-back/src/services/editor/mock-editor.ts
