+++ backend/services/pairing/pairing-back/src/utils/config.ts
@@ -4,6 +4,7 @@ dotenv.config();
 
 export default {
   RABBITMQ_URL: process.env.RABBITMQ_URL!,
+  PORT: process.env.PORT || 3000,
   LOGGING_LEVEL: process.env.LOGGING_LEVEL || "info",
   MATCHMAKING_MAX_WAIT_SECONDS: Number(process.env.MATCHMAKING_MAX_WAIT) || 100,
 };
