+++ backend/services/pairing/pairing-back/tsconfig.json
