+++ backend/services/pairing/pairing-front/.dockerignore
