+++ backend/services/pairing/pairing-front/Dockerfile
