+++ backend/services/pairing/pairing-front/src/utils/config.ts
@@ -1,0 +1,8 @@
+import dotenv from "dotenv";
+
+dotenv.config();
+
+export default {
+  RABBITMQ_URL: process.env.RABBITMQ_URL!,
+  PORT: process.env.PORT || 3000,
+};
