+++ backend/services/pairing/pairing-front/tsconfig.json
